pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "whisper-term"
version = "0.1.0"
description = "A terminal app for whisper speech to text model"
readme = "README.md"
requires-python = ">=3.8,<3.12"
authors = [
    {name = "whisper-term", email = "whisper-term@example.com"},
]
dependencies = [
    "openai-whisper>=20231117",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0,<2.0.0",
    "scipy>=1.9.0,<1.14.0",
]

[project.scripts]
whisper-term = "whisper_term.main:main"

[project.urls]
Homepage = "https://github.com/user/whisper-term"
Repository = "https://github.com/user/whisper-term"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"